+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-dev:
+1.3.1:
   - prepare sync committee duties a few epochs before change of period
   - add `controller.max-proposal-delay` option
   - integrate multi-client consensus client connections.  See configuration documentation for details.